backend/requirements.txt
fastapi
uvicorn
pymongo
httpx
python-dotenv
openai
pydantic
google-generativeai
gtts
twilio
google-cloud-texttospeech
